refactor: move common configs to shared modules
