--- shared-lib/shared-starters/starter-core/src/main/java/com/common/starter/core/exception/GlobalExceptionHandler.java
@@ -1,40 +1,0 @@
-package com.common.starter.core.exception;
-
-import com.common.dto.ErrorResponse;
-import com.common.exception.SharedException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.List;
-
-/**
- * Global exception handler providing consistent API error responses.
- */
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(SharedException.class)
-    public ResponseEntity<ErrorResponse> handleSharedException(SharedException ex) {
-        log.warn("Handled shared exception", ex);
-        ErrorResponse body = ErrorResponse.builder()
-                .code(ex.getErrorCode())
-                .message(ex.getMessage())
-                .details(ex.getDetails() == null ? null : List.of(ex.getDetails()))
-                .build();
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(body);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
-        log.error("Unhandled exception", ex);
-        ErrorResponse body = ErrorResponse.builder()
-                .code("GENERIC_ERROR")
-                .message("An unexpected error occurred")
-                .build();
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(body);
-    }
-}
