+++ shared-lib/shared-starters/starter-redis/src/main/java/com/shared/redis/starter/config/CacheAutoConfiguration.java
@@ -1,5 +1,6 @@
-package com.lms.setup.config;
+package com.shared.redis.starter.config;
 
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
@@ -17,22 +18,21 @@ import java.util.Map;
 
 @Configuration
 @EnableCaching
-public class CacheConfig {
+public class CacheAutoConfiguration {
 
     @Bean
+    @ConditionalOnMissingBean(CacheManager.class)
     public CacheManager cacheManager(RedisConnectionFactory connectionFactory) {
         Map<String, RedisCacheConfiguration> cacheConfigurations = new HashMap<>();
-        
-        // Default cache configuration with enhanced performance
+
         RedisCacheConfiguration defaultConfig = RedisCacheConfiguration.defaultCacheConfig()
                 .entryTtl(Duration.ofMinutes(30))
                 .serializeKeysWith(RedisSerializationContext.SerializationPair.fromSerializer(new StringRedisSerializer()))
                 .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(new GenericJackson2JsonRedisSerializer()))
                 .disableCachingNullValues()
-                .prefixCacheNameWith("lms:")
-                .computePrefixWith(cacheName -> "lms:" + cacheName + ":");
+                .prefixCacheNameWith("shared:")
+                .computePrefixWith(cacheName -> "shared:" + cacheName + ":");
 
-        // Specific cache configurations with optimized TTLs
         cacheConfigurations.put("countries", defaultConfig.entryTtl(Duration.ofHours(1)));
         cacheConfigurations.put("countries:active", defaultConfig.entryTtl(Duration.ofHours(2)));
         cacheConfigurations.put("cities", defaultConfig.entryTtl(Duration.ofHours(1)));
@@ -43,8 +43,6 @@ public class CacheConfig {
         cacheConfigurations.put("resources:active", defaultConfig.entryTtl(Duration.ofHours(2)));
         cacheConfigurations.put("system-parameters", defaultConfig.entryTtl(Duration.ofHours(1)));
         cacheConfigurations.put("system-parameters:active", defaultConfig.entryTtl(Duration.ofHours(2)));
-        
-        // Short-lived caches for frequently changing data
         cacheConfigurations.put("user-sessions", defaultConfig.entryTtl(Duration.ofMinutes(15)));
         cacheConfigurations.put("rate-limits", defaultConfig.entryTtl(Duration.ofMinutes(5)));
